% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{lib_versions}
\alias{lib_versions}
\title{Display versioning information of fstcore library dependencies}
\usage{
lib_versions()
}
\value{
a list with library versions
}
\description{
Display versioning information of fstcore library dependencies
}
